# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  driver: 'driver',
  runner: 'runner',
  admin: 'admin',
  unset: 'unset',
} as const;

export interface UserProfile {
  uid: string;
  email: string;
  displayName: string;
  role: UserProfileRole;
  onboardingStatus: string;
  isBanned: boolean;
  /** @nullable */
  createdAt?: string | null;
}

export interface UserActionResult {
  success: boolean;
  user?: UserProfile | null;
  /** @nullable */
  isBanned?: boolean | null;
  /** @nullable */
  onboardingStatus?: string | null;
  /** @nullable */
  message?: string | null;
}

export type DisputeStatus = typeof DisputeStatus[keyof typeof DisputeStatus];


export const DisputeStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface Dispute {
  id: string;
  gigId: string;
  reporterName: string;
  reporterRole: string;
  accusedName: string;
  reason: string;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  createdAt: string;
  /** @nullable */
  evidenceUrl?: string | null;
}

export interface DisputeResolutionInput {
  /** @minLength 1 */
  resolution: string;
}

export interface DisputeActionResult {
  success: boolean;
  dispute?: Dispute | null;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  message?: string | null;
}

export interface SkillTag {
  id: string;
  name: string;
  description: string;
  active: boolean;
}

export interface SkillTagInput {
  /** @nullable */
  id?: string | null;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  active: boolean;
}

export interface TagActionResult {
  success: boolean;
  tag?: SkillTag | null;
  /** @nullable */
  message?: string | null;
}