# lib/api-zod/src/generated/types/dispute.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DisputeStatus } from './disputeStatus';

export interface Dispute {
  id: string;
  gigId: string;
  reporterName: string;
  reporterRole: string;
  accusedName: string;
  reason: string;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  createdAt: Date;
  /** @nullable */
  evidenceUrl?: string | null;
}