# lib/api-zod/src/generated/types/disputeActionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Dispute } from './dispute';

export interface DisputeActionResult {
  success: boolean;
  dispute?: Dispute | null;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  message?: string | null;
}