# lib/api-zod/src/generated/types/disputeResolutionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DisputeResolutionInput {
  /** @minLength 1 */
  resolution: string;
}