# lib/api-zod/src/generated/types/skillTag.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillTag {
  id: string;
  name: string;
  description: string;
  active: boolean;
}