# lib/api-zod/src/generated/types/tagActionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillTag } from './skillTag';

export interface TagActionResult {
  success: boolean;
  tag?: SkillTag | null;
  /** @nullable */
  message?: string | null;
}