# lib/api-zod/src/generated/types/userActionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface UserActionResult {
  success: boolean;
  user?: UserProfile | null;
  /** @nullable */
  isBanned?: boolean | null;
  /** @nullable */
  onboardingStatus?: string | null;
  /** @nullable */
  message?: string | null;
}