# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from './userProfileRole';

export interface UserProfile {
  uid: string;
  email: string;
  displayName: string;
  role: UserProfileRole;
  onboardingStatus: string;
  isBanned: boolean;
  /** @nullable */
  createdAt?: Date | null;
}